Слияние веток

+++ src/Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
 
 
-        SalesManager oSale = new SalesManager(new int[]{});;
+        SalesManager oSale = new SalesManager(new long[]{});;
         oSale.max();
 
         System.out.println(oSale.ASale());
